export guid()
