+++ public-html/js/game.js
@@ -1,1 +1,2 @@
 console.log("test succeeded");
+console.log(document.cookie)
